Initial commit 'radix'.
This is not the first repo of its kind -- 'formulary', 'rix',
'refreshing', 'prograde', and other names have all be used before --
hopefully this will be the one that sticks, as our tooling is more
developed and our roadmap more clear than on any of the previous
embarkations.

Content from the previous iterations will slowly get folded in here as
it is updated to the new module formats; and some manual migration is
required since we're raising the bar such that everything should be
able to run in a standard way via 'stellar' rather than a blossoming of
various and ramshackle bash wrapper scripts.

And we're starting with a readme this time.  Which includes some brief
description of the standards we want to hold things to.  This could use
further fleshing-out and specifics, but we'll work towards that over
time; the heart is there.

Signed-off-by: Eric Myhre <hash@exultant.us>