Partial progress snapshot of recreating deb image.
This is a derivation from the scripts we used in the "refreshing" repo
(which is one of the ones we want to move towards archival reference
now because we've Learned Things and want to move on)... but is
incomplete and does not quite recreate the same images as previously.

I tried to break down the scripts into a few separate steps for
readability and trying to angle towards more reuse if we make more
images with different package selections... but in the process found
that some parts of the debuerreotype scripts don't compose in the way I
thought they did, so, this needs a lot more iteration time...

... and I mean *time*, because a single execution of these debootstrap
tools is just so slloooooooowwwww that it's almost comic.

Also we'll need to expose output filter configuration in stellar
modules, which is current NYI, before this will fully reproduce our old
images anyway.

So shuffling this to the back burner for a while.  Will come back later.

Signed-off-by: Eric Myhre <hash@exultant.us>