Update hello-world to include minld roosting test.
This fails, of course, but is interesting to demonstrate.

(We're still in demo mode more so than productionizing mode in this
repo for now.  In the future we'll want only working things; or,
perhaps it would be wise to develop a convention for marking "drafts"
and other demo-ish things.  We'll see.)

Signed-off-by: Eric Myhre <hash@exultant.us>