Reorg: put most things in a workspace dir.
We'll want other top level dirs in this repo over time for things like
documentation.  The "shared" dir for snippets of reusable build script
also now remains as a top leve dir.

Everything under workspace/* is now a module.

(In the future, we suppose we *could* also do this by leaving the
'.timeless' dir at the root, and configuring the workspace to map all
paths e.g. "modules/*" to ModuleName "*".  But there's very little
difference that I can think of at present, so lets do this for now.)

Only one of the modules actually changed in content, and that's because
it's using an ingest to refer to the local path to the shared scripts,
which are now one further directory up.

Signed-off-by: Eric Myhre <hash@exultant.us>