Add a smsh build.
This will be fun... using this in other formulas, we'll be able to have
arrays of commands rather than bashy stringcatting madness.

Note that there are some bumps with the present tooling though.  I'm
not sure yet if it's rio or stellar's use of git libraries, but the
ingest system doesn't understand .git *files* that redirect the object
store to another path.  So... you have to clone the repo with its
object dir in place, rather than as modern git relocates the object
dirs for submodules.  With that caveat, the build works.
Submodules are not the only approach here, nor are they a particularly
good one; we definitely don't explicitly depend on git like this in
any of the Timeless Stack tools; it just seemed like a quick solution.
In the future, I want to find find a better workflow here.

In particular, a "polination"-based approach, where we have the smsh
repo carry its own module file, and we just do coordinated releases
through the Radix workspace... should be close to ideal.  Maybe
we'll tackle that soon.  Some other workflow questions will be posed,
there, but it'll be closer to the mark.

Signed-off-by: Eric Myhre <hash@exultant.us>