Docs!
Or the attempt at docs.

There's some very skeletal todo-y bits in here.

The rough direction is what we've got right now; the details need
fleshing out (and prettier prose).

Most of the key bits at least have a brief mention now, and if it's
esoteric, hopefully it's esoteric *enough* to be searchable on the
wide web.

Signed-off-by: Eric Myhre <hash@exultant.us>