Build static m4.
Couple of things should be noted:

- Yes, significant copypasta with the gawk build.  This is
intentional; we're going to do this for a while and then see what
can be factored out (usefully) after doing a couple more projects.

- It's possible that there are some flags that aren't necessary
for this project or are redundant.  Per the above: I'm not checking,
and in fact possibly outright okay with that.  It might be simpler
to ship extra and unimpactful flags than it is useful to try to
minimize them; and that might be a valid trade for time and sanity.

Signed-off-by: Eric Myhre <hash@exultant.us>